mnlv_backend/requirements.txt
# Framework & API
django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
dj-database-url==2.1.0

# Background Tasks & Queue
celery==5.3.6
redis==5.0.3
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Download & Audio Engine
yt-dlp==2024.3.10
ffmpeg-python==0.2.0
mutagen==1.47.0

# Platform Adapters
spotipy==2.23.0
apple-music-python==1.0.6
ytmusicapi==1.6.0
tidalapi==0.7.3

# Infrastructure
gunicorn==21.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-redis==5.4.0
flower==2.0.1

# Utils
pydantic==2.7.0
requests==2.31.0
aiohttp==3.9.3
pandas
openpyxl
